Function casperwallet::key_manager::key_generator::recover_keypair

pub fn recover_keypair(secret_key_bytes: &[u8]) -> Keypair

The Keypair is stored as two seperate public and private keystrings. Inorder to maintain the functionaility of the underlying library, the keypair must be recovered when loaded from the database

Examples

let bytes = [195,172,219,111,57,44,18,178,203,227,169,238,206,170,212,61,183,60,128,214,132,158,229,243,200,89,232,38,243,15,243,70];
let keypair: Keypair = recover_keypair(&bytes);